Integer é pequeno e eu gosto de coisa longa

+++ src/main/java/com/pp1/salve/api/endereco/EnderecoController.java
@@ -1,13 +1,21 @@
 package com.pp1.salve.api.endereco;
 
-import com.pp1.salve.model.endereco.Endereco;
-import com.pp1.salve.model.endereco.EnderecoService;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.pp1.salve.model.endereco.Endereco;
+import com.pp1.salve.model.endereco.EnderecoService;
 
 @RestController
 @RequestMapping("/endereco")
@@ -23,7 +31,7 @@ public class EnderecoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Endereco> getById(@PathVariable Integer id) {
+    public ResponseEntity<Endereco> getById(@PathVariable Long id) {
         return ResponseEntity.ok(service.findById(id));
     }
 
@@ -33,13 +41,13 @@ public class EnderecoController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Endereco> update(@PathVariable Integer id, @RequestBody Endereco endereco) {
+    public ResponseEntity<Endereco> update(@PathVariable Long id, @RequestBody Endereco endereco) {
         endereco.setId(id);
         return ResponseEntity.ok(service.save(endereco));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
