+++ src/main/java/com/pp1/salve/api/item/CategoriaItemController.java
@@ -1,14 +1,22 @@
 package com.pp1.salve.api.item;
 
-import com.pp1.salve.model.item.CategoriaItem;
-import com.pp1.salve.model.item.CategoriaItemService;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.pp1.salve.model.item.CategoriaItem;
+import com.pp1.salve.model.item.CategoriaItemService;
 
 @RestController
 @RequestMapping("/categoria-item")
@@ -24,7 +32,7 @@ public class CategoriaItemController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CategoriaItem> getById(@PathVariable Integer id) {
+    public ResponseEntity<CategoriaItem> getById(@PathVariable Long id) {
         return ResponseEntity.ok(service.findById(id));
     }
 
@@ -35,13 +43,13 @@ public class CategoriaItemController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<CategoriaItem> update(@PathVariable Integer id, @RequestBody CategoriaItem categoriaItem) {
+    public ResponseEntity<CategoriaItem> update(@PathVariable Long id, @RequestBody CategoriaItem categoriaItem) {
         categoriaItem.setId(id);
         return ResponseEntity.ok(service.save(categoriaItem));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
