+++ src/main/java/com/pp1/salve/model/item/CategoriaItemRepository.java
@@ -2,6 +2,6 @@ package com.pp1.salve.model.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaItemRepository extends JpaRepository<CategoriaItem, Integer> {
+public interface CategoriaItemRepository extends JpaRepository<CategoriaItem, Long> {
   
 }
