+++ src/main/java/com/pp1/salve/model/pedido/AlteracoesPedidoRepository.java
@@ -2,6 +2,6 @@ package com.pp1.salve.model.pedido;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlteracoesPedidoRepository extends JpaRepository<AlteracoesPedido, Integer>{
+public interface AlteracoesPedidoRepository extends JpaRepository<AlteracoesPedido, Long>{
   
 }
