+++ src/main/java/com/pp1/salve/model/pedido/PedidoService.java
@@ -1,9 +1,9 @@
 package com.pp1.salve.model.pedido; 
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class PedidoService {
@@ -15,7 +15,7 @@ public class PedidoService {
         return repository.findAll();
     }
 
-    public Pedido findById(Integer id) {
+    public Pedido findById(Long id) {
         return repository.findById(id).orElseThrow(() -> new RuntimeException("Pedido não encontrado"));
     }
 
@@ -23,7 +23,7 @@ public class PedidoService {
         return repository.save(pedido);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         repository.deleteById(id);
     }
 }
